+++ apps/web/src/views/log/model/types.ts
@@ -8,32 +8,32 @@ export type CreateLogFormValues = CreateLogValues & {
   photos: PhotoPreview[];
 };
 
-export type PostTime = {
+export type Time = {
   hour: number;
   minute: number;
 };
 
-export type PostPlace = {
+export type Place = {
   name: string;
   address: string;
   latitude: number;
   longitude: number;
 };
 
-export type PostCreateRequest = {
+export type CreateRequest = {
   title: string;
   contents: string;
-  startedAt: PostTime;
-  endedAt: PostTime;
+  startedAt: Time;
+  endedAt: Time;
   studyDate: string;
   focus: number;
   scope: PostScope;
-  place: PostPlace;
+  place: Place;
   placeTags: PlaceTagValue[];
   categoryCode: string;
 };
 
-export type PostUpdateRequest = PostCreateRequest & {
+export type UpdateRequest = CreateRequest & {
   keepImageIds: number[];
 };
 
@@ -42,13 +42,13 @@ export type PostImage = {
   url: string;
 };
 
-export type PostEditFields = Omit<PostCreateRequest, 'place'> & {
+export type EditFields = Omit<CreateRequest, 'place'> & {
   studyTime: number;
-  place: PostPlace;
+  place: Place;
 };
 
-export type PostEditData = {
-  post: PostEditFields;
+export type EditData = {
+  post: EditFields;
   images: {
     images: PostImage[];
     total: number;
