+++ apps/web/src/views/log/ui/PlaceSearchOverlay.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { type ReactNode, useCallback, useEffect, useState } from 'react';
+import { type ReactNode, useCallback, useState } from 'react';
 
 import Script from 'next/script';
